refactor(ui): SP-09 and SP-11 Meshflow API client and internal_id node routes
Rename MeshtasticApi to MeshflowApi with legacy shim. Observed-node detail uses UUID
internal_id; legacy numeric URLs redirect via NodeDetails.
