+++ src/components/NodeSearch.tsx
@@ -114,7 +114,7 @@ export function NodeSearch({ onNodeSelect, displayValue, onClearSelection, lastH
               {filteredSearchResults.map((node) => (
                 <li key={node.internal_id}>
                   <Link
-                    to={onNodeSelect ? '#' : `/nodes/${node.meshtastic_node_id}`}
+                    to={onNodeSelect ? '#' : `/nodes/${node.internal_id}`}
                     className="block px-4 py-2 hover:bg-accent"
                     onClick={() => {
                       setIsOpen(false);
