+++ src/components/nodes/NodeTracerouteHistorySection.test.tsx
@@ -68,7 +68,7 @@ const mockedUseManagedNodes = vi.mocked(useManagedNodesSuspense);
 
 function makeObservedNode(overrides: Partial<ObservedNode> = {}): ObservedNode {
   return {
-    internal_id: 1,
+    internal_id: '00000000-0000-4000-8000-000000000001',
     meshtastic_node_id: 100,
     node_id_str: '!00000064',
     mac_addr: null,
