+++ src/components/nodes/RfProfileModal.test.tsx
@@ -29,7 +29,7 @@ vi.mock('@/hooks/api/useRfPropagation', () => ({
 
 function makeNode(overrides: Partial<ObservedNode> = {}): ObservedNode {
   return {
-    internal_id: 1,
+    internal_id: '00000000-0000-4000-8000-000000000001',
     meshtastic_node_id: 55,
     node_id_str: '!00000037',
     mac_addr: null,
