+++ src/components/nodes/RfPropagationSection.test.tsx
@@ -20,7 +20,7 @@ vi.mock('@/hooks/api/useRfPropagation', () => ({
 
 function makeNode(overrides: Partial<ObservedNode> = {}): ObservedNode {
   return {
-    internal_id: 1,
+    internal_id: '00000000-0000-4000-8000-000000000001',
     meshtastic_node_id: 200,
     node_id_str: '!000000c8',
     mac_addr: null,
