+++ src/components/traceroutes/TracerouteMap.test.tsx
@@ -20,7 +20,7 @@ function makeSource(): ManagedNode {
 
 function makeTarget(): ObservedNode {
   return {
-    internal_id: 2,
+    internal_id: '00000000-0000-4000-8000-000000000001',
     meshtastic_node_id: 2,
     node_id_str: '!00000002',
     mac_addr: null,
