+++ src/hooks/api/useApi.ts
@@ -1,16 +1,23 @@
 import { useMemo } from 'react';
-import { MeshtasticApi } from '@/lib/api/meshtastic-api';
+import { MeshflowApi } from '@/lib/api/meshflow-api';
 import { useConfig } from '@/providers/ConfigProvider';
+import type { AppConfig, ApiConfig } from '@/lib/types';
+
+function resolveMeshflowApiConfig(config: AppConfig): ApiConfig {
+  return config.apis.meshflow ?? config.apis.meshBot;
+}
 
 /**
- * Hook to get an instance of the MeshtasticApi
- * @returns MeshtasticApi instance
+ * Hook to get an instance of the Meshflow API client.
  */
-export function useMeshtasticApi() {
+export function useMeshflowApi() {
   const config = useConfig();
   const api = useMemo(() => {
-    return new MeshtasticApi(config.apis.meshBot);
-  }, [config.apis.meshBot]);
+    return new MeshflowApi(resolveMeshflowApiConfig(config));
+  }, [config]);
 
   return api;
 }
+
+/** @deprecated Use {@link useMeshflowApi} */
+export const useMeshtasticApi = useMeshflowApi;
