+++ src/hooks/api/useConstellationCoverage.ts
@@ -1,6 +1,6 @@
 import { keepPreviousData, useQuery } from '@tanstack/react-query';
 
-import type { ConstellationCoverageData } from '@/lib/api/meshtastic-api';
+import type { ConstellationCoverageData } from '@/lib/api/meshflow-api';
 
 import { useMeshtasticApi } from './useApi';
