+++ src/lib/infrastructure-low-battery.test.ts
@@ -10,7 +10,7 @@ import {
 
 function node(partial: Partial<ObservedNode>): ObservedNode {
   return {
-    internal_id: 1,
+    internal_id: '00000000-0000-4000-8000-000000000001',
     meshtastic_node_id: 1,
     node_id_str: '!00000001',
     mac_addr: null,
