+++ src/pages/nodes/MeshInfrastructure.tsx
@@ -323,7 +323,7 @@ function MeshInfrastructureContent() {
                         <div className="flex flex-col gap-0.5">
                           <div className="flex items-center gap-2">
                             <Link
-                              to={`/nodes/${node.meshtastic_node_id}`}
+                              to={`/nodes/${node.internal_id}`}
                               className="font-medium text-primary hover:underline"
                             >
                               {node.long_name} ({node.short_name || node.node_id_str})
@@ -384,10 +384,7 @@ function MeshInfrastructureContent() {
                       </TableCell>
                       <TableCell>
                         <div className="flex flex-col items-start gap-1 sm:flex-row sm:items-center sm:gap-3">
-                          <Link
-                            to={`/nodes/${node.meshtastic_node_id}`}
-                            className="text-primary text-sm hover:underline"
-                          >
+                          <Link to={`/nodes/${node.internal_id}`} className="text-primary text-sm hover:underline">
                             View details
                           </Link>
                           {managed != null && (
@@ -520,7 +517,7 @@ function MeshInfrastructureContent() {
                           <div className="flex flex-col gap-0.5">
                             <div className="flex items-center gap-2">
                               <Link
-                                to={`/nodes/${node.meshtastic_node_id}`}
+                                to={`/nodes/${node.internal_id}`}
                                 className="font-medium text-primary hover:underline"
                               >
                                 {node.long_name} ({node.short_name || node.node_id_str})
@@ -609,10 +606,7 @@ function MeshInfrastructureContent() {
                         </TableCell>
                         <TableCell>
                           <div className="flex flex-col items-start gap-1 sm:flex-row sm:items-center sm:gap-3">
-                            <Link
-                              to={`/nodes/${node.meshtastic_node_id}`}
-                              className="text-primary text-sm hover:underline"
-                            >
+                            <Link to={`/nodes/${node.internal_id}`} className="text-primary text-sm hover:underline">
                               View details
                             </Link>
                             {managed != null && (
