+++ src/pages/traceroutes/TriggerTracerouteModal.test.tsx
@@ -22,7 +22,7 @@ vi.mock('@/components/traceroutes/AutoTargetPreviewMap', () => ({
 
 function makeObservedNode(overrides: Partial<ObservedNode> = {}): ObservedNode {
   return {
-    internal_id: 1,
+    internal_id: '00000000-0000-4000-8000-000000000001',
     meshtastic_node_id: 42,
     node_id_str: '!0000002a',
     mac_addr: null,
