+++ src/test/test-node-ids.ts
@@ -1,0 +1,2 @@
+/** Stable UUID for observed-node fixtures in unit tests (SP-11). */
+export const TEST_OBSERVED_INTERNAL_ID = '00000000-0000-4000-8000-000000000001';
